+++ junit-docker/src/main/java/no/mnemonic/commons/junit/docker/ElasticSearchDockerResource.java
@@ -38,11 +38,16 @@ public class ElasticSearchDockerResource extends DockerResource {
 
   private final Set<String> deleteIndices;
 
-  private ElasticSearchDockerResource(String imageName, Set<Integer> applicationPort, String exposedPortsRange,
-                                      int reachabilityTimeout, Supplier<DockerClient> dockerClientResolver,
-                                      Set<String> deleteIndices, Map<String, String> environmentVariables) {
-    super(imageName, applicationPort, exposedPortsRange, reachabilityTimeout, dockerClientResolver,
-            environmentVariables);
+  private ElasticSearchDockerResource(String imageName,
+                                      Set<Integer> applicationPort,
+                                      String exposedPortsRange,
+                                      int reachabilityTimeout,
+                                      boolean skipReachabilityCheck,
+                                      Supplier<DockerClient> dockerClientResolver,
+                                      Set<String> deleteIndices,
+                                      Map<String, String> environmentVariables) {
+    super(imageName, applicationPort, exposedPortsRange, reachabilityTimeout, skipReachabilityCheck,
+            dockerClientResolver, environmentVariables);
 
     // The 'deleteIndices' parameter is optional.
     this.deleteIndices = ObjectUtils.ifNotNull(deleteIndices, Collections::unmodifiableSet, Collections.emptySet());
@@ -133,7 +138,7 @@ public class ElasticSearchDockerResource extends DockerResource {
     @Override
     public ElasticSearchDockerResource build() {
       return new ElasticSearchDockerResource(imageName, applicationPorts, exposedPortsRange, reachabilityTimeout,
-              dockerClientResolver, deleteIndices, environmentVariables);
+              skipReachabilityCheck, dockerClientResolver, deleteIndices, environmentVariables);
     }
 
     /**