+++ junit-docker/src/main/java/no/mnemonic/commons/junit/docker/MariaDBDockerResource.java
@@ -40,11 +40,17 @@ public class MariaDBDockerResource extends DockerResource {
   private final Path setupScript;
   private final Path truncateScript;
 
-  private MariaDBDockerResource(String imageName, Set<Integer> applicationPorts, String exposedPortsRange,
-                                int reachabilityTimeout, Supplier<DockerClient> dockerClientResolver,
-                                Map<String, String> environmentVariables, String setupScript, String truncateScript) {
-    super(imageName, applicationPorts, exposedPortsRange, reachabilityTimeout, dockerClientResolver,
-            environmentVariables);
+  private MariaDBDockerResource(String imageName,
+                                Set<Integer> applicationPorts,
+                                String exposedPortsRange,
+                                int reachabilityTimeout,
+                                boolean skipReachabilityCheck,
+                                Supplier<DockerClient> dockerClientResolver,
+                                Map<String, String> environmentVariables,
+                                String setupScript,
+                                String truncateScript) {
+    super(imageName, applicationPorts, exposedPortsRange, reachabilityTimeout, skipReachabilityCheck,
+            dockerClientResolver, environmentVariables);
     // Both parameters are optional.
     this.setupScript = !StringUtils.isBlank(setupScript) ? checkFileExists(setupScript) : null;
     this.truncateScript = !StringUtils.isBlank(truncateScript) ? checkFileExists(truncateScript) : null;
@@ -175,7 +181,7 @@ public class MariaDBDockerResource extends DockerResource {
       addEnvironmentVariable("MYSQL_ROOT_PASSWORD", "root");
       addEnvironmentVariable("MYSQL_ROOT_HOST", "%");
       return new MariaDBDockerResource(imageName, applicationPorts, exposedPortsRange, reachabilityTimeout,
-              dockerClientResolver, environmentVariables, setupScript, truncateScript);
+              skipReachabilityCheck, dockerClientResolver, environmentVariables, setupScript, truncateScript);
     }
 
     /**