+++ jupiter-docker/src/main/java/no/mnemonic/commons/jupiter/docker/ElasticSearchDockerExtension.java
@@ -40,10 +40,16 @@ public class ElasticSearchDockerExtension extends DockerExtension {
 
   private final Set<String> deleteIndices;
 
-  private ElasticSearchDockerExtension(String imageName, Set<Integer> applicationPort, String exposedPortsRange,
-                                       int reachabilityTimeout, Supplier<DockerClient> dockerClientResolver,
-                                       Set<String> deleteIndices, Map<String, String> environmentVariables) {
-    super(imageName, applicationPort, exposedPortsRange, reachabilityTimeout, dockerClientResolver, environmentVariables);
+  private ElasticSearchDockerExtension(String imageName,
+                                       Set<Integer> applicationPort,
+                                       String exposedPortsRange,
+                                       int reachabilityTimeout,
+                                       boolean skipReachabilityCheck,
+                                       Supplier<DockerClient> dockerClientResolver,
+                                       Set<String> deleteIndices,
+                                       Map<String, String> environmentVariables) {
+    super(imageName, applicationPort, exposedPortsRange, reachabilityTimeout, skipReachabilityCheck,
+            dockerClientResolver, environmentVariables);
 
     // The 'deleteIndices' parameter is optional.
     this.deleteIndices = ObjectUtils.ifNotNull(deleteIndices, Collections::unmodifiableSet, Collections.emptySet());
@@ -132,7 +138,7 @@ public class ElasticSearchDockerExtension extends DockerExtension {
     @Override
     public ElasticSearchDockerExtension build() {
       return new ElasticSearchDockerExtension(imageName, applicationPorts, exposedPortsRange, reachabilityTimeout,
-              dockerClientResolver, deleteIndices, environmentVariables);
+              skipReachabilityCheck, dockerClientResolver, deleteIndices, environmentVariables);
     }
 
     /**